// src/theme/types.d.ts
export declare type Breakpoints = string[];
export declare type MediaQueries = {
    xs: string;
    sm: string;
    md: string;
    lg: string;
    xl: string;
    nav: string;
};
export declare type Spacing = number[];
export declare type Radii = {
    small: string;
    default: string;
    card: string;
    circle: string;
};
export declare type Shadows = {
    level1: string;
    active: string;
    success: string;
    warning: string;
    focus: string;
    inset: string;
};
export declare type Gradients = {
    bubblegum: string;
};
export declare type Colors = {
    primary: string;
    primaryBright: string;
    primaryDark: string;
    secondary: string;
    tertiary: string;
    success: string;
    failure: string;
    warning: string;
    contrast: string;
    invertedContrast: string;
    input: string;
    inputSecondary: string;
    background: string;
    backgroundDisabled: string;
    text: string;
    textDisabled: string;
    textSubtle: string;
    borderColor: string;
    card: string;
    gradients: Gradients;
    huobi: string;
};
export declare type ZIndices = {
    dropdown: number;
    modal: number;
};